Align SemVer formatting with Git tag
The specification [1] says:

> This must follow the format of X.Y.Z or vX.Y.Z

so we can use the same format as before.

With 0c66eb2, we introduced "version" in composer.json even the usage
of the "version" property is discouraged by the specification. But it
is useful for some scripts and used as a fallback source for
rcube_plugin_api::get_info(). [2]

[1]: https://getcomposer.org/doc/04-schema.md#version
[2]: https://github.com/roundcube/roundcubemail/blob/master/program/lib/Roundcube/rcube_plugin_api.php#L265