Cleanup of class and trigger
The lookup_user_name() method was dropped, and its functionality was moved
into a code block in login_after(). The useless, corresponding
'user_create' trigger was also removed.

lookup_user_name() no longer made sense; it existed right from the beginning
because of the initial ideas taken from the new_user_identity plugin when
the development of the plugin started.